HW 11 Complete
I wish I knew how to add an image as a background or how to put text on top of an image, also I can't seem to change my font, it's always serif and Times New Roman, when I tried to change it to Arial/sans serif. My html "personal website" is officially done!